refactor dpm proof pack evidence routes
